Update logo tests in new file
